Refactor - Moved the interfaces into entities folder
