+++ src/entities/Game.ts
@@ -1,0 +1,13 @@
+import { Platform } from "./Platform";
+
+
+export interface Game {
+    id: number;
+    name: string;
+    slug: string;
+    description_raw: string;
+    background_image: string;
+    parent_platforms: { platform: Platform; }[];
+    metacritic: number;
+    rating_top: number;
+}
