+++ src/entities/Platform.ts
@@ -1,0 +1,6 @@
+
+export interface Platform {
+    id: number;
+    name: string;
+    slug: string;
+}
