+++ src/hooks/useGames.ts
@@ -1,19 +1,9 @@
 import { useInfiniteQuery } from 'react-query';
 import APIClient, { FetchResponse } from "../services/api-client";
 import useGameQueryStore from "../store";
-import { Platform } from './usePlatforms';
+import { Game } from '../entities/Game';
 
 const apiClient = new APIClient<Game>('/games');
-
-export interface Game {
-    id: number;
-    name: string;
-    slug: string;
-    description_raw: string;
-    background_image: string;
-    parent_platforms: { platform: Platform }[],
-    metacritic: number
-}
 
 const useGames = () => {
     const gameQuery = useGameQueryStore(s => s.gameQuery);
