+++ src/hooks/useGenres.tsx
@@ -2,14 +2,9 @@ import { useQuery } from "react-query";
 import genres from "../data/genres";
 import APIClient from "../services/api-client";
 import { FetchResponse } from "../services/api-client";
+import { Genre } from "../entities/Genre";
 
 const apiClient = new APIClient<Genre>("/genres");
-
-export interface Genre {
-  id: number;
-  name: string;
-  image_background: string;
-}
 
 const useGenres = () =>
   useQuery({
