+++ src/hooks/usePlatforms.ts
@@ -1,15 +1,10 @@
 import { useQuery } from "react-query";
 import APIClient, { FetchResponse } from "../services/api-client";
 import platforms from "../data/platforms";
+import { Platform } from "../entities/Platform";
 
 
 const apiClient = new APIClient<Platform>('/platforms/lists/parents');
-
-export interface Platform {
-    id: number,
-    name: string,
-    slug: string
-}
 
 const usePlatforms = () => useQuery<FetchResponse<Platform>>({
     queryKey: ['platforms'],
